requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
paho-mqtt==1.6.1
PyYAML==6.0.1
Jinja2==3.1.3
